rart

Module keys

Key types and traits for RART.

This module provides the key types and traits that can be used with Adaptive Radix Trees. Keys represent the lookup values and must implement the KeyTrait to work with the tree.

Available Key Types

ArrayKey<N>: Fixed-size keys up to N bytes, stored on the stack

VectorKey: Variable-size keys stored on the heap

Custom Keys

You can implement custom key types by implementing the KeyTrait:

use rart::keys::KeyTrait;
use rart::partials::Partial;

// Example: A wrapper around a string that implements KeyTrait
// (You would need to implement all required methods)

Modules

array_key

vector_key

Traits

KeyTrait

Trait for types that can be used as keys in an Adaptive Radix Tree.