rart::keys

Trait KeyTrait

pub trait KeyTrait:
    Clone
    + PartialEq
    + Eq
    + PartialOrd
    + Ord
    + AsRef<[u8]> {
    type PartialType: Partial + From<Self> + Clone + PartialEq;

    const MAXIMUM_SIZE: Option<usize>;

    // Required methods
    fn new_from_slice(slice: &[u8]) -> Self;
    fn new_from_partial(partial: &Self::PartialType) -> Self;
    fn extend_from_partial(&self, partial: &Self::PartialType) -> Self;
    fn truncate(&self, at_depth: usize) -> Self;
    fn at(&self, pos: usize) -> u8;
    fn length_at(&self, at_depth: usize) -> usize;
    fn to_partial(&self, at_depth: usize) -> Self::PartialType;
    fn matches_slice(&self, slice: &[u8]) -> bool;
}

Trait for types that can be used as keys in an Adaptive Radix Tree.

This trait defines the interface that all key types must implement to work with RART. Keys are the lookup values stored in the tree and must be convertible to byte sequences for trie navigation.

Requirements

Must be cloneable, comparable, and orderable

Must provide access to underlying bytes via AsRef<[u8]>

Must have an associated PartialType for prefix operations

Must support conversion to/from partials for tree operations

Example Implementation

use rart::keys::KeyTrait;
use rart::partials::{Partial, array_partial::ArrPartial};

// Note: This is a simplified example. Real implementations require
// implementing all trait methods properly.

Required Associated Constants

const MAXIMUM_SIZE: Option<usize>

Maximum size of this key type, if any.

Required Associated Types

type PartialType: Partial + From<Self> + Clone + PartialEq

The partial type associated with this key type.

Required Methods

fn new_from_slice(slice: &[u8]) -> Self

Create a new key from a byte slice.

fn new_from_partial(partial: &Self::PartialType) -> Self

Create a new key from a partial.

fn extend_from_partial(&self, partial: &Self::PartialType) -> Self

Extend this key with bytes from a partial.

fn truncate(&self, at_depth: usize) -> Self

Truncate this key to the specified depth.

fn at(&self, pos: usize) -> u8

Get the byte at the specified position.

fn length_at(&self, at_depth: usize) -> usize

Get the length of the key starting from the specified depth.

fn to_partial(&self, at_depth: usize) -> Self::PartialType

Convert part of this key to a partial starting from the specified depth.

fn matches_slice(&self, slice: &[u8]) -> bool

Check if this key matches the given byte slice exactly.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.