rart::partials

Trait Partial

pub trait Partial: AsRef<[u8]> {

    // Required methods
    fn partial_before(&self, length: usize) -> Self;
    fn partial_from(&self, src_offset: usize, length: usize) -> Self;
    fn partial_after(&self, start: usize) -> Self;
    fn partial_extended_with(&self, other: &Self) -> Self;
    fn at(&self, pos: usize) -> u8;
    fn len(&self) -> usize;
    fn prefix_length_common(&self, other: &Self) -> usize;
    fn prefix_length_key<'a, K>(&self, key: &'a K, at_depth: usize) -> usize
       where K: KeyTrait<PartialType = Self> + 'a;
    fn prefix_length_slice(&self, slice: &[u8]) -> usize;
    fn to_slice(&self) -> &[u8] ;

    // Provided methods
    fn is_empty(&self) -> bool { ... }
    fn iter(&self) -> Iter<'_, u8>  { ... }
    fn starts_with(&self, prefix: &[u8]) -> bool { ... }
    fn ends_with(&self, suffix: &[u8]) -> bool { ... }

}

Required Methods

fn partial_before(&self, length: usize) -> Self

Returns a partial up to length bytes.

fn partial_from(&self, src_offset: usize, length: usize) -> Self

Returns a partial from src_offset onwards with length bytes.

fn partial_after(&self, start: usize) -> Self

Returns a partial from start onwards.

fn partial_extended_with(&self, other: &Self) -> Self

Extends the partial with another partial.

fn at(&self, pos: usize) -> u8

Returns the byte at pos.

fn len(&self) -> usize

Returns the length of the partial.

fn prefix_length_common(&self, other: &Self) -> usize

Returns the length of the common prefix between self and other.

fn prefix_length_key<'a, K>(&self, key: &'a K, at_depth: usize) -> usize
where K: KeyTrait<PartialType = Self> + 'a,

Returns the length of the common prefix between self and key.

fn prefix_length_slice(&self, slice: &[u8]) -> usize

Returns the length of the common prefix between self and slice.

fn to_slice(&self) -> &[u8]

Return a slice form of the partial. Warning: could take copy, depending on the implementation. Really just for debugging purposes.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the partial is empty.

fn iter(&self) -> Iter<'_, u8>

Returns an iterator over the bytes in the partial.

fn starts_with(&self, prefix: &[u8]) -> bool

Returns true if the partial starts with the given prefix.

fn ends_with(&self, suffix: &[u8]) -> bool

Returns true if the partial ends with the given suffix.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.